Sample Data

The Bokeh.Data module provides access to some small sample datasets

See the Python Bokeh reference for more details of these datasets.

Unless otherwise stated, all these functions return a Tables.jl-compatible table. Optionally a materializer can be specified to convert it to some other type (such as rowtable, columntable or DataFrame).

Bokeh.Data.anscombe — Function

anscombe([materializer])

The four data series that comprise 'Anscombe's Quartet'.

Bokeh.Data.antibiotics — Function

antibiotics([materializer])

A table of Will Burtin's historical data regarding antibiotic efficacies.

Bokeh.Data.autompg — Function

autompg([materializer])

A version of the Auto MPG data set.

Bokeh.Data.autompg_clean — Function

autompg_clean([materializer])

A version of the Auto MPG data set. Cleans the mfr and origin fields.

Bokeh.Data.autompg2 — Function

autompg2([materializer])

A version of the Auto MPG data set.

Bokeh.Data.browsers_nov_2013 — Function

browsers_nov_2013([materializer])

Browser market share by version from November 2013.

Bokeh.Data.commits — Function

commits([materializer])

Time series of commits for a GitHub user between 2012 and 2016.

Bokeh.Data.daylight_warsaw_2013 — Function

daylight_warsaw_2013([materializer])

2013 Warsaw daylight hours.

Bokeh.Data.degrees — Function

degrees([materializer])

A table of data regarding bachelor's degrees earned by women.

Bokeh.Data.elements — Function

elements([materializer])

A periodic table dataset.

Bokeh.Data.iris — Function

iris([materializer])

Fisher's Iris dataset.

Bokeh.Data.les_mis — Function

les_mis([materializer])

Co-occurrence of characters in Les Miserables.

Returns a named tuple (nodes=..., links=...) of nodes (characters) and links (co-occurrences) between them.

Bokeh.Data.numberly — Function

numberly([materializer])

The numberly dataset.

Bokeh.Data.obiszow_mtb_xcm — Function

obiszow_mtb_xcm([materializer])

Route data (including altitude) for a bike race in Eastern Europe.

Bokeh.Data.olympics2014 — Function

olympics2014([materializer])

Medal counts by country for the 2014 Olympics.

Bokeh.Data.penguins — Function

penguins([materializer])

The Palmer Archipelago (Antarctica) penguin dataset.

Bokeh.Data.probly — Function

probly([materializer])

The probly dataset.

Bokeh.Data.sea_surface_temperature — Function

sea_surface_temperature([materializer])

Time series of historical average sea surface temperatures.

Bokeh.Data.sprint — Function

sprint([materializer])

Historical results for Olympic sprints by year.

Bokeh.Data.unemployment1948 — Function

unemployment1948([materializer])

US Unemployment rate data by month and year, from 1948 to 2013.

Bokeh.Data.us_holidays — Function

us_holidays([materializer])

Calendar file of US Holidays from Mozilla provided by icalendar.

Bokeh.Data.us_marriages_divorces — Function

us_marriages_divorces([materializer])

U.S. marriage and divorce statistics between 1867 and 2014.

Bokeh.Data.us_states — Function

us_states([materializer])

Geometry data for US States.