Interaction

Bokeh.js_link — Function

js_link(model, property, other_model, other_property)

Link two Bokeh model properties via JavaScript.

Whenever model.property is changed, then other_model.other_property is set to its value.

Bokeh.js_on_click — Function

js_on_click(model, callbacks...)

Attach CustomJS callbacks to run when the model is clicked.

This applies to button widgets such as Button, Toggle and Dropdown. It also includes group widgets such as CheckboxGroup, RadioGroup, CheckboxButtonGroup and RadioButtonGroup.

Bokeh.js_on_change — Function

js_on_change(model, event, callbacks...)

Attach CustomJS callbacks to an arbitrary BokehJS model change event.

Change events for model properties are of the form "change:property_name" but as a convenience you can simply provide "property_name".

Bokeh.js_on_event — Function

js_on_event(model, event, callbacks...)

Attach CustomJS callbacks to an arbitrary BokehJS model event.