Settings & Backends

Bokeh.setting — Function

setting(name::Symbol)

Retrieve the setting with the given name.

See settings! for the possible settings.

Bokeh.settings! — Function

settings!(key=value, ...)

Update the global settings for Bokeh.

offline: If set to true then any generated plots will work offline. Default: false.

browser_cmd: The command used to open the browser. Default: Operating-system dependent.

BrowserDisplay

Bokeh.BrowserDisplay — Type

BrowserDisplay()

Displays Bokeh plots in the Browser.

Activate with pushdisplay(Bokeh.BrowserDisplay()).

BokehBlink

The BokehBlink package can be separately installed to enable plotting into a standalone Blink.jl window.

BokehBlink.Display — Type

BokehBlink.Display()

Displays Bokeh plots in a Blink window.

Activate with pushdisplay(BokehBlink.Display()).

BokehBlink.save — Function

save(io_or_filename, [plot]; mime="image/png")

Save a screenshot of the current plot or the given plot.