Plotting

Basics

Bokeh.figure — Function

figure(; ...)

Create a new Figure and return it.

Acceptable keyword arguments are:

Anything taken by Figure.

x_range/y_range: Sets the x/y-range. May be a vector of factors or a 2-tuple representing an interval. Default: DataRange1d().

x_axis/y_axis: Sets the x/y-axis. May be nothing to suppress. Default: LinearAxis().

x_axis_location/y_axis_location: Where to put the axis. One of "left", "right", "above" or "below". Default: "below"/"left".

x_axis_label/y_axis_label: Sets the label on the x/y-axis.

x_grid/y_grid: Sets the x/y-grid. May be nothing to suppress.

tools: Optional list of tools to create a toolbar from.

tooltips: If given, add a HoverTool with these tooltips.

Bokeh.plot! — Function

plot!(plot, item; ...)
plot!(plot, type; ...)

Adds a new item to the plot, or an item of the given type.

When passing a type, the allowed keyword arguments include anything accepted by the type. Some additional arguments are allowed depending on what item or type is.

The constructed item is returned. In the case of glyphs, the corresponding GlyphRenderer is returned instead.

Glyphs

Additional keyword arguments:

Anything accepted by GlyphRenderer.

source: Alias for data_source. May be a DataSource, Dict of columns, or a Tables.jl-style table.

color: Alias for all the *_color properties.

alpha: Alias for all the *_alpha properties.

palette: If the glyph has a color_mapper property, it is set to a LinearColorMapper with this palette.

legend_label

legend_field

legend_group

filters: List of filters to apply to the source data.

Renderers

This includes axes, grids, legends and other annotations.

Additional keyword arguments:

location: One of "center" (default), "left", "right", "below" or "above".

dimension: For axes, you must specify either the location or dimension and the other one is inferred.

Tools

Additional keyword arguments:

activate: If true, then set the tool as the active one of its kind on the toolbar.

Layouts

Bokeh.column — Function

column(items; ...)
column(items...; ...)

Create a new Column with the given items.

Bokeh.row — Function

row(items; ...)
row(items...; ...)

Create a new Row with the given items.

Transforms

These can be used as the value for properties such as fill_color or mark.

Bokeh.dodge — Function

dodge(field, value; kw...)

Transform the given field with a Dodge.

Bokeh.factor_cmap — Function

factor_cmap(field, palette, factors; ...)

Transform the given field with a CategoricalColorMapper that selects the given colors for the corresponding factors.

Bokeh.factor_mark — Function

factor_mark(field, markers, factors; ...)

Transform the given field with a CategoricalMarkerMapper that selects the given markers for the corresponding factors.

Bokeh.factor_hatch — Function

factor_hatch(field, patterns, factors; ...)

Transform the given field with a CategoricalPatternMapper that selects the given hatch patterns for the corresponding factors.

Bokeh.jitter — Function

jitter(field, width; ...)

Transform the given field with a Jitter that applies random pertubations of the given width.

Bokeh.linear_cmap — Function

linear_cmap(field, palette; ...)

Transform the given field with a LinearColorMapper that selects colors linearly from the given palette.

Bokeh.log_cmap — Function

log_cmap(field, palette; ...)

Transform the given field with a LogColorMapper that selects colors logarithmically from the given palette.

Bokeh.transform — Function

transform(field, transform)

A the field of the given name transformed with the given Transform.