Settings & Backends

Bokeh.settings! — Function

settings!(key=value, ...)

Update the global settings for Bokeh.

display: The display backend to use, such as :browser or :blink. Default: :null.

offline: If set to true then any generated plots will work offline. Default: false.

browser_cmd: The command used to open the browser. Default: Operating-system dependent.

Display Backends

Plots can automatically be displayed anywhere supporting HTML output, such as a Pluto or Jupyter notebook. Anywhere else, such as the Julia REPL, will require a display backend to be activated.

Browser

To display plots in your default browser, set

Bokeh.settings!(display=:browser)

Blink

The BokehBlink package can be separately installed to enable plotting into a standalone Blink.jl window.

To activate this backend, do

using Bokeh, BokehBlink
Bokeh.settings!(display=:blink)

This package can also be used to export plots as images. You do not need to activate the backend first.

BokehBlink.save — Function

save(io, [plot]; ...)
save(filename, [plot]; ...)

Save a screenshot of the current plot or the given plot.

Keyword Arguments

format: The output format to use. Valid formats include "png", "jpeg" and "webp". If filename is given, the format is inferred from the extension. Otherwise it defaults to "png".

quality: A number between 0 and 1 specifying the quality (compression level) of the image.

Warning

This is an experimental feature and likely to be buggy, particularly the version taking a plot argument.

Null

The null backend cannot display anything. It is the default backend. Use it to deactivate another backend:

Bokeh.settings!(display=:null)

Interactive Backends

🚧 Coming soon! 🚧