Plotting

Figure

Bokeh.figure — Function

figure(; ...)

Create a new Figure and return it.

Acceptable keyword arguments are:

Anything taken by Figure.

x_range/y_range: Sets the x/y-range. May be a vector of factors or a 2-tuple representing an interval.

x_axis_type/y_axis_type: The type of axis. One of "linear", "log", "datetime", "mercator", "auto" or nothing (for no axis).

x_axis_location/y_axis_location: Where to put the axis. One of "left", "right", "above" or "below".

x_axis_label/y_axis_label: The label to show on the axis.

tools: Optional list of tools to create a toolbar from.

tooltips: If given, add a HoverTool with these tooltips.

Glyphs

Bokeh.annular_wedge! — Function

annular_wedge!(plot; kw...)

Adds a AnnularWedge to the given plot.

If you do not specify any source data, then a data source is created from all the data arguments which are vectors.

Keyword arguments

Anything accepted by AnnularWedge.

Anything accepted by GlyphRenderer.

source: Shorthand for data_source. May be a ColumnDataSource, Dict or Tables.jl-compatible table.

color: Shorthand for all *_color properties.

alpha: Shorthand for all *_alpha properties.

palette: Sets a color mapper with this palette.

legend_label: Adds an item to the legend with the given label.

legend_field: Adds items to the legend from the given field.

legend_group: Adds items to the legend grouping on the given field.

filters: A vector of filters to select a subset of the data.

Bokeh.annulus! — Function

annulus!(plot; kw...)

Adds a Annulus to the given plot.

If you do not specify any source data, then a data source is created from all the data arguments which are vectors.

Keyword arguments

Anything accepted by Annulus.

Anything accepted by GlyphRenderer.

source: Shorthand for data_source. May be a ColumnDataSource, Dict or Tables.jl-compatible table.

color: Shorthand for all *_color properties.

alpha: Shorthand for all *_alpha properties.

palette: Sets a color mapper with this palette.

legend_label: Adds an item to the legend with the given label.

legend_field: Adds items to the legend from the given field.

legend_group: Adds items to the legend grouping on the given field.

filters: A vector of filters to select a subset of the data.

Bokeh.arc! — Function

arc!(plot; kw...)

Adds a Arc to the given plot.

If you do not specify any source data, then a data source is created from all the data arguments which are vectors.

Keyword arguments

Anything accepted by Arc.

Anything accepted by GlyphRenderer.

source: Shorthand for data_source. May be a ColumnDataSource, Dict or Tables.jl-compatible table.

color: Shorthand for all *_color properties.

alpha: Shorthand for all *_alpha properties.

palette: Sets a color mapper with this palette.

legend_label: Adds an item to the legend with the given label.

legend_field: Adds items to the legend from the given field.

legend_group: Adds items to the legend grouping on the given field.

filters: A vector of filters to select a subset of the data.

Bokeh.bezier! — Function

bezier!(plot; kw...)

Adds a Bezier to the given plot.

If you do not specify any source data, then a data source is created from all the data arguments which are vectors.

Keyword arguments

Anything accepted by Bezier.

Anything accepted by GlyphRenderer.

source: Shorthand for data_source. May be a ColumnDataSource, Dict or Tables.jl-compatible table.

color: Shorthand for all *_color properties.

alpha: Shorthand for all *_alpha properties.

palette: Sets a color mapper with this palette.

legend_label: Adds an item to the legend with the given label.

legend_field: Adds items to the legend from the given field.

legend_group: Adds items to the legend grouping on the given field.

filters: A vector of filters to select a subset of the data.

Bokeh.circle! — Function

circle!(plot; kw...)

Adds a Circle to the given plot.

If you do not specify any source data, then a data source is created from all the data arguments which are vectors.

Keyword arguments

Anything accepted by Circle.

Anything accepted by GlyphRenderer.

source: Shorthand for data_source. May be a ColumnDataSource, Dict or Tables.jl-compatible table.

color: Shorthand for all *_color properties.

alpha: Shorthand for all *_alpha properties.

palette: Sets a color mapper with this palette.

legend_label: Adds an item to the legend with the given label.

legend_field: Adds items to the legend from the given field.

legend_group: Adds items to the legend grouping on the given field.

filters: A vector of filters to select a subset of the data.

Bokeh.ellipse! — Function

ellipse!(plot; kw...)

Adds a Ellipse to the given plot.

If you do not specify any source data, then a data source is created from all the data arguments which are vectors.

Keyword arguments

Anything accepted by Ellipse.

Anything accepted by GlyphRenderer.

source: Shorthand for data_source. May be a ColumnDataSource, Dict or Tables.jl-compatible table.

color: Shorthand for all *_color properties.

alpha: Shorthand for all *_alpha properties.

palette: Sets a color mapper with this palette.

legend_label: Adds an item to the legend with the given label.

legend_field: Adds items to the legend from the given field.

legend_group: Adds items to the legend grouping on the given field.

filters: A vector of filters to select a subset of the data.

Bokeh.harea! — Function

harea!(plot; kw...)

Adds a HArea to the given plot.

If you do not specify any source data, then a data source is created from all the data arguments which are vectors.

Keyword arguments

Anything accepted by HArea.

Anything accepted by GlyphRenderer.

source: Shorthand for data_source. May be a ColumnDataSource, Dict or Tables.jl-compatible table.

color: Shorthand for all *_color properties.

alpha: Shorthand for all *_alpha properties.

palette: Sets a color mapper with this palette.

legend_label: Adds an item to the legend with the given label.

legend_field: Adds items to the legend from the given field.

legend_group: Adds items to the legend grouping on the given field.

filters: A vector of filters to select a subset of the data.

Bokeh.hbar! — Function

hbar!(plot; kw...)

Adds a HBar to the given plot.

If you do not specify any source data, then a data source is created from all the data arguments which are vectors.

Keyword arguments

Anything accepted by HBar.

Anything accepted by GlyphRenderer.

source: Shorthand for data_source. May be a ColumnDataSource, Dict or Tables.jl-compatible table.

color: Shorthand for all *_color properties.

alpha: Shorthand for all *_alpha properties.

palette: Sets a color mapper with this palette.

legend_label: Adds an item to the legend with the given label.

legend_field: Adds items to the legend from the given field.

legend_group: Adds items to the legend grouping on the given field.

filters: A vector of filters to select a subset of the data.

Bokeh.hextile! — Function

hextile!(plot; kw...)

Adds a HexTile to the given plot.

If you do not specify any source data, then a data source is created from all the data arguments which are vectors.

Keyword arguments

Anything accepted by HexTile.

Anything accepted by GlyphRenderer.

source: Shorthand for data_source. May be a ColumnDataSource, Dict or Tables.jl-compatible table.

color: Shorthand for all *_color properties.

alpha: Shorthand for all *_alpha properties.

palette: Sets a color mapper with this palette.

legend_label: Adds an item to the legend with the given label.

legend_field: Adds items to the legend from the given field.

legend_group: Adds items to the legend grouping on the given field.

filters: A vector of filters to select a subset of the data.

Bokeh.image! — Function

image!(plot; kw...)

Adds a Image to the given plot.

If you do not specify any source data, then a data source is created from all the data arguments which are vectors.

Keyword arguments

Anything accepted by Image.

Anything accepted by GlyphRenderer.

source: Shorthand for data_source. May be a ColumnDataSource, Dict or Tables.jl-compatible table.

color: Shorthand for all *_color properties.

alpha: Shorthand for all *_alpha properties.

palette: Sets a color mapper with this palette.

legend_label: Adds an item to the legend with the given label.

legend_field: Adds items to the legend from the given field.

legend_group: Adds items to the legend grouping on the given field.

filters: A vector of filters to select a subset of the data.

Bokeh.image_rgba! — Function

image_rgba!(plot; kw...)

Adds a ImageRGBA to the given plot.

If you do not specify any source data, then a data source is created from all the data arguments which are vectors.

Keyword arguments

Anything accepted by ImageRGBA.

Anything accepted by GlyphRenderer.

source: Shorthand for data_source. May be a ColumnDataSource, Dict or Tables.jl-compatible table.

color: Shorthand for all *_color properties.

alpha: Shorthand for all *_alpha properties.

palette: Sets a color mapper with this palette.

legend_label: Adds an item to the legend with the given label.

legend_field: Adds items to the legend from the given field.

legend_group: Adds items to the legend grouping on the given field.

filters: A vector of filters to select a subset of the data.

Bokeh.image_url! — Function

image_url!(plot; kw...)

Adds a ImageURL to the given plot.

If you do not specify any source data, then a data source is created from all the data arguments which are vectors.

Keyword arguments

Anything accepted by ImageURL.

Anything accepted by GlyphRenderer.

source: Shorthand for data_source. May be a ColumnDataSource, Dict or Tables.jl-compatible table.

color: Shorthand for all *_color properties.

alpha: Shorthand for all *_alpha properties.

palette: Sets a color mapper with this palette.

legend_label: Adds an item to the legend with the given label.

legend_field: Adds items to the legend from the given field.

legend_group: Adds items to the legend grouping on the given field.

filters: A vector of filters to select a subset of the data.

Bokeh.line! — Function

line!(plot; kw...)

Adds a Line to the given plot.

If you do not specify any source data, then a data source is created from all the data arguments which are vectors.

Keyword arguments

Anything accepted by Line.

Anything accepted by GlyphRenderer.

source: Shorthand for data_source. May be a ColumnDataSource, Dict or Tables.jl-compatible table.

color: Shorthand for all *_color properties.

alpha: Shorthand for all *_alpha properties.

palette: Sets a color mapper with this palette.

legend_label: Adds an item to the legend with the given label.

legend_field: Adds items to the legend from the given field.

legend_group: Adds items to the legend grouping on the given field.

filters: A vector of filters to select a subset of the data.

Bokeh.multi_line! — Function

multi_line!(plot; kw...)

Adds a MultiLine to the given plot.

If you do not specify any source data, then a data source is created from all the data arguments which are vectors.

Keyword arguments

Anything accepted by MultiLine.

Anything accepted by GlyphRenderer.

source: Shorthand for data_source. May be a ColumnDataSource, Dict or Tables.jl-compatible table.

color: Shorthand for all *_color properties.

alpha: Shorthand for all *_alpha properties.

palette: Sets a color mapper with this palette.

legend_label: Adds an item to the legend with the given label.

legend_field: Adds items to the legend from the given field.

legend_group: Adds items to the legend grouping on the given field.

filters: A vector of filters to select a subset of the data.

Bokeh.multi_polygons! — Function

multi_polygons!(plot; kw...)

Adds a MultiPolygons to the given plot.

If you do not specify any source data, then a data source is created from all the data arguments which are vectors.

Keyword arguments

Anything accepted by MultiPolygons.

Anything accepted by GlyphRenderer.

source: Shorthand for data_source. May be a ColumnDataSource, Dict or Tables.jl-compatible table.

color: Shorthand for all *_color properties.

alpha: Shorthand for all *_alpha properties.

palette: Sets a color mapper with this palette.

legend_label: Adds an item to the legend with the given label.

legend_field: Adds items to the legend from the given field.

legend_group: Adds items to the legend grouping on the given field.

filters: A vector of filters to select a subset of the data.

Bokeh.oval! — Function

oval!(plot; kw...)

Adds a Oval to the given plot.

If you do not specify any source data, then a data source is created from all the data arguments which are vectors.

Keyword arguments

Anything accepted by Oval.

Anything accepted by GlyphRenderer.

source: Shorthand for data_source. May be a ColumnDataSource, Dict or Tables.jl-compatible table.

color: Shorthand for all *_color properties.

alpha: Shorthand for all *_alpha properties.

palette: Sets a color mapper with this palette.

legend_label: Adds an item to the legend with the given label.

legend_field: Adds items to the legend from the given field.

legend_group: Adds items to the legend grouping on the given field.

filters: A vector of filters to select a subset of the data.

Bokeh.patch! — Function

patch!(plot; kw...)

Adds a Patch to the given plot.

If you do not specify any source data, then a data source is created from all the data arguments which are vectors.

Keyword arguments

Anything accepted by Patch.

Anything accepted by GlyphRenderer.

source: Shorthand for data_source. May be a ColumnDataSource, Dict or Tables.jl-compatible table.

color: Shorthand for all *_color properties.

alpha: Shorthand for all *_alpha properties.

palette: Sets a color mapper with this palette.

legend_label: Adds an item to the legend with the given label.

legend_field: Adds items to the legend from the given field.

legend_group: Adds items to the legend grouping on the given field.

filters: A vector of filters to select a subset of the data.

Bokeh.patches! — Function

patches!(plot; kw...)

Adds a Patches to the given plot.

If you do not specify any source data, then a data source is created from all the data arguments which are vectors.

Keyword arguments

Anything accepted by Patches.

Anything accepted by GlyphRenderer.

source: Shorthand for data_source. May be a ColumnDataSource, Dict or Tables.jl-compatible table.

color: Shorthand for all *_color properties.

alpha: Shorthand for all *_alpha properties.

palette: Sets a color mapper with this palette.

legend_label: Adds an item to the legend with the given label.

legend_field: Adds items to the legend from the given field.

legend_group: Adds items to the legend grouping on the given field.

filters: A vector of filters to select a subset of the data.

Bokeh.quad! — Function

quad!(plot; kw...)

Adds a Quad to the given plot.

If you do not specify any source data, then a data source is created from all the data arguments which are vectors.

Keyword arguments

Anything accepted by Quad.

Anything accepted by GlyphRenderer.

source: Shorthand for data_source. May be a ColumnDataSource, Dict or Tables.jl-compatible table.

color: Shorthand for all *_color properties.

alpha: Shorthand for all *_alpha properties.

palette: Sets a color mapper with this palette.

legend_label: Adds an item to the legend with the given label.

legend_field: Adds items to the legend from the given field.

legend_group: Adds items to the legend grouping on the given field.

filters: A vector of filters to select a subset of the data.

Bokeh.quadratic! — Function

quadratic!(plot; kw...)

Adds a Quadratic to the given plot.

If you do not specify any source data, then a data source is created from all the data arguments which are vectors.

Keyword arguments

Anything accepted by Quadratic.

Anything accepted by GlyphRenderer.

source: Shorthand for data_source. May be a ColumnDataSource, Dict or Tables.jl-compatible table.

color: Shorthand for all *_color properties.

alpha: Shorthand for all *_alpha properties.

palette: Sets a color mapper with this palette.

legend_label: Adds an item to the legend with the given label.

legend_field: Adds items to the legend from the given field.

legend_group: Adds items to the legend grouping on the given field.

filters: A vector of filters to select a subset of the data.

Bokeh.ray! — Function

ray!(plot; kw...)

Adds a Ray to the given plot.

If you do not specify any source data, then a data source is created from all the data arguments which are vectors.

Keyword arguments

Anything accepted by Ray.

Anything accepted by GlyphRenderer.

source: Shorthand for data_source. May be a ColumnDataSource, Dict or Tables.jl-compatible table.

color: Shorthand for all *_color properties.

alpha: Shorthand for all *_alpha properties.

palette: Sets a color mapper with this palette.

legend_label: Adds an item to the legend with the given label.

legend_field: Adds items to the legend from the given field.

legend_group: Adds items to the legend grouping on the given field.

filters: A vector of filters to select a subset of the data.

Bokeh.rect! — Function

rect!(plot; kw...)

Adds a Rect to the given plot.

If you do not specify any source data, then a data source is created from all the data arguments which are vectors.

Keyword arguments

Anything accepted by Rect.

Anything accepted by GlyphRenderer.

source: Shorthand for data_source. May be a ColumnDataSource, Dict or Tables.jl-compatible table.

color: Shorthand for all *_color properties.

alpha: Shorthand for all *_alpha properties.

palette: Sets a color mapper with this palette.

legend_label: Adds an item to the legend with the given label.

legend_field: Adds items to the legend from the given field.

legend_group: Adds items to the legend grouping on the given field.

filters: A vector of filters to select a subset of the data.

Bokeh.scatter! — Function

scatter!(plot; kw...)

Adds a Scatter to the given plot.

If you do not specify any source data, then a data source is created from all the data arguments which are vectors.

Keyword arguments

Anything accepted by Scatter.

Anything accepted by GlyphRenderer.

source: Shorthand for data_source. May be a ColumnDataSource, Dict or Tables.jl-compatible table.

color: Shorthand for all *_color properties.

alpha: Shorthand for all *_alpha properties.

palette: Sets a color mapper with this palette.

legend_label: Adds an item to the legend with the given label.

legend_field: Adds items to the legend from the given field.

legend_group: Adds items to the legend grouping on the given field.

filters: A vector of filters to select a subset of the data.

Bokeh.segment! — Function

segment!(plot; kw...)

Adds a Segment to the given plot.

If you do not specify any source data, then a data source is created from all the data arguments which are vectors.

Keyword arguments

Anything accepted by Segment.

Anything accepted by GlyphRenderer.

source: Shorthand for data_source. May be a ColumnDataSource, Dict or Tables.jl-compatible table.

color: Shorthand for all *_color properties.

alpha: Shorthand for all *_alpha properties.

palette: Sets a color mapper with this palette.

legend_label: Adds an item to the legend with the given label.

legend_field: Adds items to the legend from the given field.

legend_group: Adds items to the legend grouping on the given field.

filters: A vector of filters to select a subset of the data.

Bokeh.step! — Function

step!(plot; kw...)

Adds a Step to the given plot.

If you do not specify any source data, then a data source is created from all the data arguments which are vectors.

Keyword arguments

Anything accepted by Step.

Anything accepted by GlyphRenderer.

source: Shorthand for data_source. May be a ColumnDataSource, Dict or Tables.jl-compatible table.

color: Shorthand for all *_color properties.

alpha: Shorthand for all *_alpha properties.

palette: Sets a color mapper with this palette.

legend_label: Adds an item to the legend with the given label.

legend_field: Adds items to the legend from the given field.

legend_group: Adds items to the legend grouping on the given field.

filters: A vector of filters to select a subset of the data.

Bokeh.text! — Function

text!(plot; kw...)

Adds a Text to the given plot.

If you do not specify any source data, then a data source is created from all the data arguments which are vectors.

Keyword arguments

Anything accepted by Text.

Anything accepted by GlyphRenderer.

source: Shorthand for data_source. May be a ColumnDataSource, Dict or Tables.jl-compatible table.

color: Shorthand for all *_color properties.

alpha: Shorthand for all *_alpha properties.

palette: Sets a color mapper with this palette.

legend_label: Adds an item to the legend with the given label.

legend_field: Adds items to the legend from the given field.

legend_group: Adds items to the legend grouping on the given field.

filters: A vector of filters to select a subset of the data.

Bokeh.varea! — Function

varea!(plot; kw...)

Adds a VArea to the given plot.

If you do not specify any source data, then a data source is created from all the data arguments which are vectors.

Keyword arguments

Anything accepted by VArea.

Anything accepted by GlyphRenderer.

source: Shorthand for data_source. May be a ColumnDataSource, Dict or Tables.jl-compatible table.

color: Shorthand for all *_color properties.

alpha: Shorthand for all *_alpha properties.

palette: Sets a color mapper with this palette.

legend_label: Adds an item to the legend with the given label.

legend_field: Adds items to the legend from the given field.

legend_group: Adds items to the legend grouping on the given field.

filters: A vector of filters to select a subset of the data.

Bokeh.vbar! — Function

vbar!(plot; kw...)

Adds a VBar to the given plot.

If you do not specify any source data, then a data source is created from all the data arguments which are vectors.

Keyword arguments

Anything accepted by VBar.

Anything accepted by GlyphRenderer.

source: Shorthand for data_source. May be a ColumnDataSource, Dict or Tables.jl-compatible table.

color: Shorthand for all *_color properties.

alpha: Shorthand for all *_alpha properties.

palette: Sets a color mapper with this palette.

legend_label: Adds an item to the legend with the given label.

legend_field: Adds items to the legend from the given field.

legend_group: Adds items to the legend grouping on the given field.

filters: A vector of filters to select a subset of the data.

Bokeh.wedge! — Function

wedge!(plot; kw...)

Adds a Wedge to the given plot.

If you do not specify any source data, then a data source is created from all the data arguments which are vectors.

Keyword arguments

Anything accepted by Wedge.

Anything accepted by GlyphRenderer.

source: Shorthand for data_source. May be a ColumnDataSource, Dict or Tables.jl-compatible table.

color: Shorthand for all *_color properties.

alpha: Shorthand for all *_alpha properties.

palette: Sets a color mapper with this palette.

legend_label: Adds an item to the legend with the given label.

legend_field: Adds items to the legend from the given field.

legend_group: Adds items to the legend grouping on the given field.

filters: A vector of filters to select a subset of the data.

Axes & Grids

Bokeh.grid! — Function

grid!(plot; axis, dimension, ...)

Add a Grid to the given plot and return it.

You must specify the axis and dimension (0 for x, 1 for y) it relates to.

Bokeh.linear_axis! — Function

linear_axis!(plot; location, ...)

Add a LinearAxis to the given plot and return it.

The location must be one of "left", "right", "above" or "below".

Bokeh.log_axis! — Function

log_axis!(plot; location, ...)

Add a LogAxis to the given plot and return it.

The location must be one of "left", "right", "above" or "below".

Bokeh.categorical_axis! — Function

categorical_axis!(plot; location, ...)

Add a CategoricalAxis to the given plot and return it.

The location must be one of "left", "right", "above" or "below".

Bokeh.datetime_axis! — Function

datetime_axis!(plot; location, ...)

Add a DatetimeAxis to the given plot and return it.

The location must be one of "left", "right", "above" or "below".

Bokeh.mercator_axis! — Function

mercator_axis!(plot; location, ...)

Add a MercatorAxis to the given plot and return it.

The location must be one of "left", "right", "above" or "below".

Tools

Bokeh.pan_tool! — Function

pan_tool!(plot; active=false, ...)

Add a PanTool to the given plot and return it.

If active=true then the tool becomes the active one of its kind.

Bokeh.range_tool! — Function

range_tool!(plot; active=false, ...)

Add a RangeTool to the given plot and return it.

If active=true then the tool becomes the active one of its kind.

Bokeh.wheelpan_tool! — Function

wheelpan_tool!(plot; active=false, ...)

Add a WheelPanTool to the given plot and return it.

If active=true then the tool becomes the active one of its kind.

Bokeh.wheelzoom_tool! — Function

wheelzoom_tool!(plot; active=false, ...)

Add a WheelZoomTool to the given plot and return it.

If active=true then the tool becomes the active one of its kind.

Bokeh.save_tool! — Function

save_tool!(plot; active=false, ...)

Add a SaveTool to the given plot and return it.

If active=true then the tool becomes the active one of its kind.

Bokeh.reset_tool! — Function

reset_tool!(plot; active=false, ...)

Add a ResetTool to the given plot and return it.

If active=true then the tool becomes the active one of its kind.

Bokeh.tap_tool! — Function

tap_tool!(plot; active=false, ...)

Add a TapTool to the given plot and return it.

If active=true then the tool becomes the active one of its kind.

Bokeh.crosshair_tool! — Function

crosshair_tool!(plot; active=false, ...)

Add a CrosshairTool to the given plot and return it.

If active=true then the tool becomes the active one of its kind.

Bokeh.boxzoom_tool! — Function

boxzoom_tool!(plot; active=false, ...)

Add a BoxZoomTool to the given plot and return it.

If active=true then the tool becomes the active one of its kind.

Bokeh.zoomin_tool! — Function

zoomin_tool!(plot; active=false, ...)

Add a ZoomInTool to the given plot and return it.

If active=true then the tool becomes the active one of its kind.

Bokeh.zoomout_tool! — Function

zoomout_tool!(plot; active=false, ...)

Add a ZoomOutTool to the given plot and return it.

If active=true then the tool becomes the active one of its kind.

Bokeh.boxselect_tool! — Function

boxselect_tool!(plot; active=false, ...)

Add a BoxSelectTool to the given plot and return it.

If active=true then the tool becomes the active one of its kind.

Bokeh.lassoselect_tool! — Function

lassoselect_tool!(plot; active=false, ...)

Add a LassoSelectTool to the given plot and return it.

If active=true then the tool becomes the active one of its kind.

Bokeh.polyselect_tool! — Function

polyselect_tool!(plot; active=false, ...)

Add a PolySelectTool to the given plot and return it.

If active=true then the tool becomes the active one of its kind.

Bokeh.help_tool! — Function

help_tool!(plot; active=false, ...)

Add a HelpTool to the given plot and return it.

If active=true then the tool becomes the active one of its kind.

Layouts

Bokeh.column — Function

column(items; ...)
column(items...; ...)

Create a new Column with the given items.

Bokeh.row — Function

row(items; ...)
row(items...; ...)

Create a new Row with the given items.

Transforms

These can be used as the value for properties such as fill_color or mark.

Bokeh.dodge — Function

dodge(field, value; kw...)

Transform the given field with a Dodge.

Bokeh.factor_cmap — Function

factor_cmap(field, palette, factors; ...)

Transform the given field with a CategoricalColorMapper that selects the given colors for the corresponding factors.

Bokeh.factor_mark — Function

factor_mark(field, markers, factors; ...)

Transform the given field with a CategoricalMarkerMapper that selects the given markers for the corresponding factors.

Bokeh.factor_hatch — Function

factor_hatch(field, patterns, factors; ...)

Transform the given field with a CategoricalPatternMapper that selects the given hatch patterns for the corresponding factors.

Bokeh.jitter — Function

jitter(field, width; ...)

Transform the given field with a Jitter that applies random pertubations of the given width.

Bokeh.linear_cmap — Function

linear_cmap(field, palette; ...)

Transform the given field with a LinearColorMapper that selects colors linearly from the given palette.

Bokeh.log_cmap — Function

log_cmap(field, palette; ...)

Transform the given field with a LogColorMapper that selects colors logarithmically from the given palette.

Bokeh.transform — Function

transform(field, transform)

A the field of the given name transformed with the given Transform.